# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.CreateMemberInput:
    properties:
      email:
        type: string
      full_name:
        type: string
      phone:
        type: string
      sponsor_id:
        type: string
    required:
    - full_name
    type: object
...
  models.MemberResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      joined_at:
        type: string
      member_code:
        type: string
      phone:
        type: string
      sponsor_id:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  models.MemberWithDownlineCount:
    properties:
      created_at:
        type: string
      downline_count:
        type: integer
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      joined_at:
        type: string
      level_1_count:
        type: integer
      level_2_count:
        type: integer
      level_3_count:
        type: integer
      member_code:
        type: string
      phone:
        type: string
      sponsor_id:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  models.UpdateMemberInput:
    properties:
      email:
        type: string
      full_name:
        type: string
      phone:
        type: string
      status:
        enum:
        - active
        - inactive
        - pending
        - suspended
        type: string
    type: object
...
  utils.Meta:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  utils.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      meta:
        $ref: '#/definitions/utils.Meta'
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: |-
    Admin API for Rudra management system

    This API provides endpoints for managing Rudra operations including:
    - Admin authentication and authorization
    - Member management
    - Referral tree management
    - Income calculations and distributions
    - Reports and analytics
    - Audit logging
  title: Rudra Admin API
  version: 1.0.0
paths:
...
  /api/v1/members:
    get:
      consumes:
      - application/json
      description: Get paginated list of members with filtering
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by status
        enum:
        - active
        - inactive
        - pending
        - suspended
        in: query
        name: status
        type: string
      - description: Search by name, code, email, or phone
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.MemberResponse'
                  type: array
              type: object
      security:
      - Bearer: []
      summary: List members
      tags:
      - members
    post:
      consumes:
      - application/json
      description: Create a new member in the Rudra system
      parameters:
      - description: Member creation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateMemberInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.MemberResponse'
              type: object
      security:
      - Bearer: []
      summary: Create member
      tags:
      - members
  /api/v1/members/{id}:
    delete:
      description: Soft delete a member (cannot delete members with downlines)
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - Bearer: []
      summary: Delete member
      tags:
      - members
    get:
      description: Get a single member by their ID
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.MemberResponse'
              type: object
      security:
      - Bearer: []
      summary: Get member by ID
      tags:
      - members
    put:
      consumes:
      - application/json
      description: Update an existing member's details
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: string
      - description: Member update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateMemberInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.MemberResponse'
              type: object
      security:
      - Bearer: []
      summary: Update member
      tags:
      - members
  /api/v1/members/{id}/downline:
    get:
      description: Get the downline tree for a member
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: string
      - default: 10
        description: Max depth level
        in: query
        name: level
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.MemberWithDownlineCount'
                  type: array
              type: object
      security:
      - Bearer: []
      summary: Get member downline tree
      tags:
      - members
  /api/v1/members/{id}/upline:
    get:
      description: Get the upline chain for a member
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: string
      - default: 10
        description: Max depth level
        in: query
        name: level
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.MemberResponse'
                  type: array
              type: object
      security:
      - Bearer: []
      summary: Get member upline
      tags:
      - members
  /api/v1/members/search:
    get:
      description: Search members by keyword
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.MemberResponse'
                  type: array
              type: object
      security:
      - Bearer: []
      summary: Search members
      tags:
      - members
  /api/v1/members/stats:
    get:
      description: Get member statistics (total, active, inactive counts)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - Bearer: []
      summary: Get member stats
      tags:
      - members
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"